[ADD] command CD and PWD

+++ srcs/server_cmd_get.c
@@ -6,7 +6,7 @@
 /*   By: agrumbac <agrumbac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/18 19:45:12 by agrumbac          #+#    #+#             */
-/*   Updated: 2018/11/20 22:58:37 by agrumbac         ###   ########.fr       */
+/*   Updated: 2018/11/20 23:26:20 by agrumbac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ bool			cmd_get(int sock, uint64_t body_size)
 	char			*filename;
 	void			*file;
 	size_t			file_size;
-	int				ret;
+	ssize_t			ret;
 
 	if (body_size > MAXPATHLEN)
 		return(cmd_bad(sock, ERR_PATHLEN_OVERFLOW));
